requirements.txt
# altair==6.0.0
# annotated-types==0.7.0
# anyio==4.12.1
# attrs==25.4.0
# blinker==1.9.0
# cachetools==6.2.6
# certifi==2026.2.25
# cffi==2.0.0
# charset-normalizer==3.4.4
# click==8.3.1
# colorama==0.4.6
# contourpy==1.3.3
# cryptography==46.0.5
# cycler==0.12.1
# distro==1.9.0
# fonttools==4.61.1
# gitdb==4.0.12
# GitPython==3.1.46
# google-ai-generativelanguage==0.6.15
# google-api-core==2.30.0
# google-api-python-client==2.191.0
# google-auth==2.48.0
# google-auth-httplib2==0.3.0
# google-genai==1.65.0
# googleapis-common-protos==1.72.0
# groq==1.0.0
# grpcio==1.78.0
# grpcio-status==1.71.2
# h11==0.16.0
# httpcore==1.0.9
# httplib2==0.31.2
# httpx==0.28.1
# idna==3.11
# Jinja2==3.1.6
# jsonschema==4.26.0
# jsonschema-specifications==2025.9.1
# kiwisolver==1.4.9
# MarkupSafe==3.0.3
# matplotlib==3.10.8
# narwhals==2.17.0
# numpy==2.4.2
# packaging==26.0
# pandas==2.3.3
# pillow==12.1.1
# proto-plus==1.27.1
# protobuf==5.29.6
# pyarrow==23.0.1
# pyasn1==0.6.2
# pyasn1_modules==0.4.2
# pycparser==3.0
# pydantic==2.12.5
# pydantic_core==2.41.5
# pydeck==0.9.1
# pyparsing==3.3.2
# python-dateutil==2.9.0.post0
# python-dotenv==1.2.2
# pytz==2025.2
# referencing==0.37.0
# requests==2.32.5
# rpds-py==0.30.0
# rsa==4.9.1
# seaborn==0.13.2
# six==1.17.0
# smmap==5.0.2
# sniffio==1.3.1
# streamlit==1.54.0
# tenacity==9.1.4
# toml==0.10.2
# tornado==6.5.4
# tqdm==4.67.3
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# tzdata==2025.3
# uritemplate==4.2.0
# urllib3==2.6.3
# watchdog==6.0.0
# websockets==16.0





streamlit==1.54.0
pandas==2.3.3
numpy==2.4.2
matplotlib==3.10.8
seaborn==0.13.2
python-dotenv==1.2.2
groq==1.0.0
altair==6.0.0
pydeck==0.9.1
watchdog==6.0.0